Changed usize to u64/u32 on serde derived structs
In all the locations where a struct derives Serialize or Deserialize usize has been replaced
with either u64 or u32. No instances of isize existed. In the cases where code using the
data required a usize u32 was chosen as the replacement. There is no point in avoiding usize
in the on disk formats if 32-bit platforms can't read and use the values written.

+++ actor/src/lib.rs
@@ -288,7 +288,7 @@ where
 #[async_trait::async_trait]
 pub trait Actor: Send + Sized + 'static {
     /// Capacity of actor queue
-    fn mailbox_capacity(&self) -> usize {
+    fn mailbox_capacity(&self) -> u32 {
         100
     }
 
@@ -347,8 +347,8 @@ pub struct InitializedActor<A: Actor> {
 
 impl<A: Actor> InitializedActor<A> {
     /// Constructor.
-    pub fn new(actor: A, mailbox_capacity: usize) -> Self {
-        let (sender, receiver) = mpsc::channel(mailbox_capacity);
+    pub fn new(actor: A, mailbox_capacity: u32) -> Self {
+        let (sender, receiver) = mpsc::channel(mailbox_capacity as usize);
         InitializedActor {
             actor,
             address: Addr::new(sender),
