+++ core/src/sumeragi/mod.rs
@@ -202,9 +202,9 @@ where
     /// [`iroha_p2p::Network`] actor address
     pub network: Addr<IrohaNetwork>,
     /// Mailbox size
-    pub mailbox: usize,
+    pub mailbox: u32,
     fault_injection: PhantomData<F>,
-    gossip_batch_size: usize,
+    gossip_batch_size: u32,
     gossip_period: Duration,
 }
 
@@ -317,7 +317,7 @@ pub const TELEMETRY_INTERVAL: Duration = Duration::from_secs(5);
 impl<G: GenesisNetworkTrait, K: KuraTrait, W: WorldTrait, F: FaultInjection> Actor
     for SumeragiWithFault<G, K, W, F>
 {
-    fn mailbox_capacity(&self) -> usize {
+    fn mailbox_capacity(&self) -> u32 {
         self.mailbox
     }
 
@@ -1777,9 +1777,9 @@ pub mod config {
     const DEFAULT_COMMIT_TIME_MS: u64 = 2000;
     const DEFAULT_TX_RECEIPT_TIME_MS: u64 = 500;
     const DEFAULT_N_TOPOLOGY_SHIFTS_BEFORE_RESHUFFLE: u64 = 1;
-    const DEFAULT_MAILBOX_SIZE: usize = 100;
+    const DEFAULT_MAILBOX_SIZE: u32 = 100;
     const DEFAULT_GOSSIP_PERIOD_MS: u64 = 1000;
-    const DEFAULT_GOSSIP_BATCH_SIZE: usize = 500;
+    const DEFAULT_GOSSIP_BATCH_SIZE: u32 = 500;
 
     /// `SumeragiConfiguration` provides an ability to define parameters such as `BLOCK_TIME_MS`
     /// and list of `TRUSTED_PEERS`.
@@ -1806,9 +1806,9 @@ pub mod config {
         /// Limits to which transactions must adhere
         pub transaction_limits: TransactionLimits,
         /// Mailbox size
-        pub mailbox: usize,
+        pub mailbox: u32,
         /// Maximum number of transactions in tx gossip batch message. While configuring this, attention should be payed to `p2p` max message size.
-        pub gossip_batch_size: usize,
+        pub gossip_batch_size: u32,
         /// Period in milliseconds for pending transaction gossiping between peers.
         pub gossip_period_ms: u64,
     }
